+++ CHANGELOG.md
@@ -24,7 +24,7 @@
   * **Breaking**:
     * `Commerce_Customer_Address`: rename field `StreetNr` to `StreetNumber`, `lastname` to `lastName` and `firstname` to `firstName` 
     * `Commerce_Customer_Result`: `defaultShippingAddress` and `defaultBillingAddress` now can return null if there is no default address
-    * `Commerce_Customer_PersonData`: field `birthday` is now nullable. 
+    * `Commerce_Customer_PersonData`: field `birthday` is now nullable and of type `Date`. 
 
 **payment**
 * Introduced wallet payment method